requirements.txt
llama-index
python-dotenv
atproto
twitter-openapi-python
beautifulsoup4
feedparser
sqlalchemy
schedule
websockets 